requirements.txt
torch
Pillow
florence2
transformers
opencv-python
pyyaml
python-dotenv
openai